# io_buffer.rb, request.rb - improve handling with a body array/enumeration