@MSP-Greg 2021-09-14
PR #2595 (08-Apr) included code to test performance with three types of response bodies, array, chunked, and string.  Quite a few code optimizations later, I've ran tests on the code here.

The code was run with `-w2 -t5:5`, using an abstract UNIXSocket, which isn't available with wrk.  The speed increase varies, but is significant with larger array or chunked bodies.

Note that the code uses files from both #2694 and #2695, so I'll wait for those...

<details>
  <summary>Speed comparison PR vs Master</summary>

This PR:
```
Request / Response time distribution in mS
10000 requests - 10 loops of 100 clients * 10 requests per client

───────────────────────────────────────────────────────────────────────────────────    1kB Body
       req/sec   10%    20%    40%    50%    60%    80%    90%    95%    97%    99%
 array   9823  0.300  0.330  0.386  0.425  0.551  0.852  0.994  1.209  1.356  1.742
 chunk   9902  0.293  0.325  0.388  0.458  0.629  0.865  0.995  1.187  1.320  1.618
string   9909  0.294  0.326  0.384  0.434  0.582  0.862  1.007  1.226  1.364  1.678

───────────────────────────────────────────────────────────────────────────────────   10kB Body
       req/sec   10%    20%    40%    50%    60%    80%    90%    95%    97%    99%
 array   9650  0.314  0.346  0.399  0.434  0.509  0.851  1.011  1.238  1.381  1.716
 chunk   9559  0.310  0.345  0.408  0.465  0.619  0.886  1.022  1.222  1.362  1.674
string   9696  0.307  0.341  0.400  0.443  0.556  0.866  1.011  1.216  1.368  1.664

───────────────────────────────────────────────────────────────────────────────────  100kB Body
       req/sec   10%    20%    40%    50%    60%    80%    90%    95%    97%    99%
 array   6185  0.830  0.890  0.980  1.020  1.070  1.330  1.570  1.840  2.020  2.800
 chunk   5754  0.920  0.990  1.070  1.110  1.180  1.520  1.720  1.960  2.110  2.460
string   6254  0.830  0.900  0.980  1.020  1.070  1.390  1.600  1.860  2.020  2.490

300 requests - 10 loops of 15 clients * 2 requests per client
─────────────────────────────────────────────────────────────────────────────────── 2050kB Body
       req/sec   10%    20%    40%    50%    60%    80%    90%    95%    97%    99%
 array     98   87.3   89.9   92.9   94.3   96.2  100.8  107.9  123.8  131.8  142.4
 chunk     98   86.6   89.9   94.7   96.6   98.2  103.2  107.1  112.5  117.9  222.8
string    111   77.5   81.0   84.3   85.6   87.4   90.7   94.4   96.7   98.0  102.9
```

Master:
```
Request / Response time distribution in mS
10000 requests - 10 loops of 100 clients * 10 requests per client

───────────────────────────────────────────────────────────────────────────────────    1kB Body
       req/sec   10%    20%    40%    50%    60%    80%    90%    95%    97%    99%
 array   8722  0.306  0.349  0.571  0.726  0.816  0.973  1.185  1.391  1.554  1.941
 chunk   7284  0.327  0.494  0.810  0.898  1.074  1.384  1.639  1.852  1.965  2.243
string   9427  0.302  0.337  0.413  0.541  0.712  0.902  1.045  1.251  1.395  1.643

───────────────────────────────────────────────────────────────────────────────────   10kB Body
       req/sec   10%    20%    40%    50%    60%    80%    90%    95%    97%    99%
 array   6590  0.340  0.430  0.890  1.090  1.280  1.590  1.840  2.040  2.210  2.570
 chunk   3963  0.724  1.287  1.893  2.149  2.375  2.959  3.416  3.800  4.043  4.604
string   9261  0.314  0.349  0.424  0.539  0.711  0.917  1.067  1.284  1.424  1.707

───────────────────────────────────────────────────────────────────────────────────  100kB Body
       req/sec   10%    20%    40%    50%    60%    80%    90%    95%    97%    99%
 array   1744   2.96   3.93   4.96   5.37   5.79   6.70   7.40   8.01   8.41   9.11
 chunk    600  12.40  13.70  15.50  16.20  17.00  18.80  20.10  21.20  21.90  23.30
string   6150   0.84   0.90   0.98   1.02   1.09   1.46   1.64   1.89   2.05   2.44

300 requests - 10 loops of 15 clients * 2 requests per client
─────────────────────────────────────────────────────────────────────────────────── 2050kB Body
       req/sec   10%    20%    40%    50%    60%    80%    90%    95%    97%    99%
 array     74  117.7  121.9  126.1  127.9  130.5  135.7  139.4  142.2  146.7  157.9
 chunk     31  274.4  298.5  311.9  316.8  320.9  330.5  339.5  346.4  350.9  371.5
string    117   74.2   76.8   79.9   81.3   82.9   86.2   89.1   90.8   92.9   94.3
```

</details>