@MSP-Greg 2022-01-01
I've been working on all these files, adding comments, renaming methods, fixing some bad initial decisions for parameters, etc.

Closing, almost finished with updates.  Happy New Year!